requirements.txt
# requirements.txt
tensorflow==2.11.0
numpy==1.21.6
matplotlib==3.5.1
pillow==9.3.0
scikit-image==0.19.2
scikit-learn==1.0.2
